Refactor development
